requirements.txt
fastapi
uvicorn
numpy==1.24.4
Pillow
open-clip-torch==2.20.0
faiss-cpu
opencv-python
matplotlib
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
python-multipart